// src/models/item.model.ts
export interface ItemModel {
  id: number;
  itemName: string;
  itemRarity: string;
  itemCost: number;
  itemType: string;
  itemRestrictions: string;
  itemAttunement: string;
  itemSource: string;
  itemUrl: string;
  itemVisualDesc: string;
  itemShopkeeperDesc: string;
  itemSelected: boolean;
}